+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unrealeased)
 
+- Add support for using the before_all hook with Rails' parallelize feature (using processes). ([@peret][])
+
+Make sure to include `TestProf::BeforeAll::Minitest` before you call `parallelize`.
+
 ## 1.0.6 (2021-06-23)
 
 - Fix Spring detection when `DISABLE_SPRING=1` is used. ([@palkan][])